add support for specifying what to include in response to source
Added 'fieldIds', a new parameter to 'response()', an array of
dot-notation parameter ids that defines what lead data will be
included in the response back to the inbound source. This is
being added for use by a new Lendio integration, which will
use this module but specify other data to include in the response.